Implement Redis client connect pool metrics
This adds an OpenCensus dependency, and is only really used
if there's a valid exporter compiled in your code. Values
are collected from the Redis client's PoolStats() function at
an interval specified in seconds in the environment
variable metrics_recording_interval_seconds. If the
environment variable is not specified or parses to 0, then
no gathering of data from connection pools will be done.